requirements.txt
alembic==1.3.2
astroid==2.3.3
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pep8==1.7.1
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
